refactor code best practinces
